Follow up #407: float64 validation reductions, swap leftovers, portable oracles (#411)
Rework the validation reductions in float64, scaling each column by a power of two. Inputs whose span, or whose cancellation, exceeds binary64 resolution are refused instead of being reported as a wrong number. Remove leftovers from the #407 type swap (bit-identical), warm the float64 exponent scan directly, and replace the longdouble test references with portable exact references plus a guard against extended dtypes.

The QP repair-row change was reverted after a complete-fit comparison with master. It and the sum_to_zero underflow check are follow-ups. Native, real-data, development and security CI passed, and the Opus 5.5 and Codex reviews found nothing blocking.

+++ src/superglm/distributional/results/solver.py
@@ -437,9 +437,7 @@ def _assessment_objective_error_bound(
         )
     except ValueError as evidence:
         raise ValueError(f"{failure_context}: invalid original evidence") from evidence
-    return float(
-        _enclosed_bound_sum(scalar, np.float64(0.5) * error, np.float64(0.5) * original_error)
-    )
+    return float(_enclosed_bound_sum(scalar, 0.5 * error, 0.5 * original_error))
 
 
 def _assessment_face_geometry_matches(left: PenaltyFace, right: PenaltyFace) -> bool:
